requirements.txt
aiohttp-3.9.5
asyncpraw-7.7.1
openai-1.37.1
praw-7.7.1
python-dotenv-1.0.1
yt-dlp==2024.7.25
